Skip all WS between JSON vals to minimise levels of recursive calls.
Although the recursive call to next() seems to be tail-recursive, some
of the fuzzer inputs were large enough for there to be more than 10,000
levels of recursion at which point the program segfaulted
catastrophically (ie, no ASAN output; nothing readable in gdb; no "stack
overflow" message anywhere). Skipping all whitespace before the
recursive call minimises the number of levels of recursion because it
ensures that the next thing in the stream is a JSON token in which case
that event is returned immediately.

In the process also fixed some incorrect separator logic and its
associated test case.